service/requirements.txt
Django==4.0
djangorestframework==3.14.0
celery[redis]==5.2.7
flower==1.2.0
celery_singleton==0.3.1
django-redis==5.2.0
django-cachalot==2.5.2
gunicorn==21.2.0
pre-commit==3.3.3
black==23.7.0
autoflake==2.2.0
pyupgrade==3.10.1
reorder-python-imports==3.10.0
flake8==6.1.0
pytest-django==4.5.2
pytest-cov==4.1.0
requests==2.31.0
atlassian-python-api==3.41.1
django-storages==1.14